+++ src/main/java/pro/sky/springcalc/CalcService.java
@@ -5,5 +5,5 @@ public interface CalcService {
     String calculatorPlus(Integer num1, Integer num2);
     String calculatorMinus(Integer num1, Integer num2);
     String calculatorMultiply(Integer num1, Integer num2);
-    String calculatorDivide(Double num1, Double num2);
+    String calculatorDivide(Integer num1, Integer num2);
 }
